Use Python script to sync navbar and footer
The Jekyll includes didn't work for some reason, and
GitHub Pages doesn't support Server-Side Includes, but
I like using them to test out my site on Apache, so
I wrote a tiny script to literally copy the includes.
The actual site files are in the site folder, and
running make.py adds the includes in the root folder.
Finally, .htaccess and .nojekyll are back to stay.